Add Job Statuses

+++ src/api/DBManager.cpp
@@ -1,5 +1,5 @@
 /**
- * @file DBReader.cpp
+ * @file DBManager.cpp
  * @author Ryan Butler
  * @brief Implements the DB Reader class
  * @version 0.1
@@ -9,75 +9,73 @@
  * 
  */
 
-#include "api/DBReader.hpp"
+#include "api/DBManager.hpp"
 #include <format>
 #include <iostream>
 #include <unordered_map>
 #include <cstdlib> 
 
-DBReader::DBReader(bool testDB){
+DBManager::DBManager(bool testDB){
     if (testDB){
-        init("host=localhost port=5432 dbname=trafficDBTest");
+        connectionStr_ = "host=localhost port=5432 dbname=trafficDBTest";
     } else {
-        init("host=localhost port=5432 dbname=trafficDB");
+        connectionStr_ = "host=localhost port=5432 dbname=trafficDB";
     }
 }
 
-void DBReader::init(std::string connectionStr){
+std::expected<pqxx::connection, std::string> DBManager::getConnection(){
     try {   
-        connect_ = std::make_shared<pqxx::connection>(connectionStr);
-        // Initialize tables if not initialized
-
-        /// @warning This code is duplicated from the DBLoggerBase's setup code
-        // Create 3 tables: Traffic Jobs Main Table, Car Data Table (including strategies), Main data table (with car x,v,t values)
-        pqxx::work tx(*connect_);
-        tx.exec("CREATE TABLE IF NOT EXISTS trafficJobs ( jobID int GENERATED ALWAYS AS IDENTITY PRIMARY KEY, configfile text, jobname text)");
-        tx.exec("CREATE TABLE IF NOT EXISTS carData (carID int, jobID int, follow text, lead text, FOREIGN KEY (jobID) REFERENCES trafficjobs(jobID) , PRIMARY KEY (carID, jobID))");
-        tx.exec("CREATE TABLE IF NOT EXISTS snapshotData (carID int, jobID int, x float, v float, t float, PRIMARY KEY (carID, jobID, t), FOREIGN KEY (carID, jobID) REFERENCES cardata (carID, jobID))");
-        tx.commit();
+        return pqxx::connection(connectionStr_);
     } catch(const std::exception& e) {
         std::cout << "Error connecting to the database" << std::endl;
         std::cerr << e.what() << '\n';
-    }
-    if (!connect_){
-        throw DBReadError("Fatal Error: Could not connect to the database!");
+        return std::unexpected(std::format("Error Connecting to the database: {}", e.what()));
     }
 }
 
+
 // JOB DATA
 
-std::expected<JobData, std::string> DBReader::queryJobs(std::string jobname){
-    std::string querystr = std::format("SELECT jobname, configfile FROM TrafficJobs WHERE jobname = '{}'", jobname);
-    pqxx::work tx{*connect_};
+std::expected<JobData, std::string> DBManager::queryJobs(std::string jobname){
+    std::string querystr = std::format("SELECT jobname, configfile, status, error  FROM TrafficJobs WHERE jobname = '{}'", jobname);
+
+    auto connect = getConnection();
+    if (!connect){return std::unexpected(connect.error());}
+    pqxx::work tx{*connect};
     pqxx::result result;
     try {
         result = tx.exec(querystr);
     } catch(const std::exception& e) {
         return std::unexpected(std::format("Error executing SELECT query: {}",  e.what()));
     }
-    
+
     if (result.empty()) {
         return std::unexpected("No job named " + jobname);
     }
     pqxx::row r = result[0];
-    std::string name,cfgfile;
+    std::string name,cfgfile,error, status;
     try {
-        name = r[0].as<std::string>();
-        cfgfile = r[1].as<std::string>();
+        name = r["jobname"].as<std::string>();
+        cfgfile = r["configfile"].as<std::string>();
+        error = r["error"].as<std::string>();
+        status = r["status"].as<std::string>();
+
     } catch(const std::exception& e) {
-        return std::unexpected("Error converting name or cfgfile to a string");
+        return std::unexpected("Error converting name, cfgfile, error or status to a string");
     }
-    
-    return JobData{name, cfgfile};
+
+    return JobData{name, cfgfile, error, status};
 }
 
-std::expected<std::vector<JobData>, std::string> DBReader::queryJobs(){
-    std::string querystr = std::format("SELECT jobname, configfile FROM TrafficJobs");
-    pqxx::work tx{*connect_};
+std::expected<std::vector<JobData>, std::string> DBManager::queryJobs(){
+    std::string querystr = std::format("SELECT jobname, configfile, status, error FROM TrafficJobs");
+    auto connect = getConnection();
+    if (!connect){return std::unexpected(connect.error());}
+    pqxx::work tx{*connect};
     std::vector<JobData> data;
     try {
-        for (auto [name, cfg] : tx.query<std::string, std::string>(querystr)){
-            data.push_back({name, cfg});
+        for (auto [name, cfg, status, error] : tx.query<std::string, std::string, std::string, std::string>(querystr)){
+            data.push_back({name, cfg, error, status});
         }    
     } catch(const std::exception& e) {
         return std::unexpected(std::format("Error in SELECT query from DB Reader: ", e.what()));
@@ -88,8 +86,10 @@ std::expected<std::vector<JobData>, std::string> DBReader::queryJobs(){
 
 // CAR METADATA
 
-std::expected<CarMetadata, std::string>  DBReader::queryCars(std::string jobname, int carid){
-    pqxx::work tx{*connect_};
+std::expected<CarMetadata, std::string>  DBManager::queryCars(std::string jobname, int carid){
+    auto connect = getConnection();
+    if (!connect){return std::unexpected(connect.error());}
+    pqxx::work tx{*connect};
 
     std::string querystr = std::format("SELECT follow, lead FROM CarData INNER JOIN TrafficJobs ON TrafficJobs.JobID = CarData.JobID WHERE jobname = '{}' and carid = {}", jobname, carid);
     pqxx::result result = tx.exec(querystr);
@@ -100,8 +100,10 @@ std::expected<CarMetadata, std::string>  DBReader::queryCars(std::string jobname
     return CarMetadata{r[0].as<std::string>(), r[1].as<std::string>(), {}, carid};
 }
 
-std::expected<std::vector<CarMetadata>, std::string> DBReader::queryCars(std::string jobname){
-    pqxx::work tx{*connect_};
+std::expected<std::vector<CarMetadata>, std::string> DBManager::queryCars(std::string jobname){
+    auto connect = getConnection();
+    if (!connect){return std::unexpected(connect.error());}
+    pqxx::work tx{*connect};
     std::string querystr = std::format("SELECT carid, follow, lead FROM CarData INNER JOIN TrafficJobs ON TrafficJobs.JobID = CarData.JobID WHERE jobname = '{}'", jobname);
 
     std::vector<CarMetadata> data;
@@ -122,9 +124,12 @@ std::expected<std::vector<CarMetadata>, std::string> DBReader::queryCars(std::st
 
 // RAW DATA 
 
-std::expected<RawData, std::string>  DBReader::queryData(std::string jobname, int carid){
+std::expected<RawData, std::string>  DBManager::queryData(std::string jobname, int carid){
     std::string querystr = std::format("SELECT x, v, t FROM snapshotData INNER JOIN TrafficJobs ON TrafficJobs.JobID = snapshotData.jobid WHERE TrafficJobs.jobname = '{}' and snapshotData.carid = {} ORDER BY snapshotData.t ASC", jobname, carid);
-    pqxx::work tx{*connect_};
+    auto connect = getConnection();
+    if (!connect){return std::unexpected(connect.error());}
+    pqxx::work tx{*connect};
+    
     RawData raw;
     raw.id_ = carid;
     try {
@@ -140,11 +145,13 @@ std::expected<RawData, std::string>  DBReader::queryData(std::string jobname, in
     return raw;
 }
 
-std::expected<std::vector<RawData>, std::string> DBReader::queryData(std::string jobname){
+std::expected<std::vector<RawData>, std::string> DBManager::queryData(std::string jobname){
 
     std::string querystr = std::format("SELECT carid, x, v, t FROM snapshotData INNER JOIN TrafficJobs ON TrafficJobs.JobID = snapshotData.jobid WHERE TrafficJobs.jobname = '{}' ORDER BY snapshotData.carid ASC, snapshotData.t ASC", jobname);
 
-    pqxx::work tx{*connect_};
+    auto connect = getConnection();
+    if (!connect){return std::unexpected(connect.error());}
+    pqxx::work tx{*connect};
 
     // Holds a vector of snapshots for each car
     std::vector<RawData> alldata;
@@ -167,8 +174,11 @@ std::expected<std::vector<RawData>, std::string> DBReader::queryData(std::string
     return alldata;
 }
 
-std::expected<std::vector<int>, std::string> DBReader::getJobId(std::string jobname) {
-    pqxx::work tx{*connect_};
+std::expected<std::vector<int>, std::string> DBManager::getJobId(std::string jobname) {
+    
+    auto connect = getConnection();
+    if (!connect){return std::unexpected(connect.error());}
+    pqxx::work tx{*connect};
     std::vector<int> ids;
     std::string querystr = std::format("SELECT jobID FROM TrafficJobs WHERE jobname = '{}'", jobname);
     try {
@@ -185,12 +195,14 @@ std::expected<std::vector<int>, std::string> DBReader::getJobId(std::string jobn
     }
 }
 
-std::expected<void, std::string> DBReader::deleteJob(std::string jobname){
+DBResponse DBManager::deleteJob(std::string jobname){
 
-    auto deleteJobByID = [this, &jobname](const std::vector<int>&  jobids) -> std::expected<void, std::string>{
+    auto deleteJobByID = [this, &jobname](const std::vector<int>&  jobids) -> DBResponse {
         try {
             for (const int& id : jobids){
-                pqxx::work tx{*connect_};
+                auto connect = getConnection();
+                if (!connect){return std::unexpected(connect.error());}
+                pqxx::work tx{*connect};
                 std::string querystr = std::format("DELETE FROM snapshotData  WHERE jobid = '{}'", id);
                 tx.exec(querystr);
                 querystr = std::format("DELETE FROM CarData WHERE jobid = '{}'", id);
@@ -201,7 +213,7 @@ std::expected<void, std::string> DBReader::deleteJob(std::string jobname){
                 tx.exec(querystr);
                 tx.commit();
             }
-            return std::expected<void, std::string>{};
+            return DBResponse{};
         } catch(const std::exception& e) {
             std::string errMsg = std::format("Error deleting {} from database: {}", jobname, e.what());
             return std::unexpected(errMsg);
